fix: Toss 결제 상태 매핑 보완

+++ backend/src/main/java/com/team08/backend/domain/payment/dto/toss/TossPaymentResponse.java
@@ -7,7 +7,7 @@ public record TossPaymentResponse(
         String orderId,
         String status,
         String method,
-        int totalAmount,
+        long totalAmount,
         OffsetDateTime approvedAt
 ) {
 }
